requirements.txt
asgiref==3.8.1
contourpy==1.3.2
cycler==0.12.1
Django==5.2
django-js-asset==3.1.2
django-tinymce==4.1.0
fonttools==4.57.0
kiwisolver==1.4.8
matplotlib==3.10.1
mysqlclient==2.2.7
narwhals==1.37.1
numpy==2.2.5
packaging==25.0
pandas==2.2.3
pillow==11.2.1
plotly==6.0.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
